renamed Dice to die as it is gramatically correct
