+++ dicechanics/__init__.py
@@ -1,4 +1,4 @@
-from dicechanics._Dice import Dice
+from dicechanics._Die import Die
 from dicechanics._Pool import Pool
 from dicechanics._interface import d, z, pool
 from dicechanics._standard import d4, d6, d8, d10, d12, d20, d100, z9
