+++ dicechanics/typing.py
@@ -1,5 +1,6 @@
 from typing import Callable, Any
 
+
 type NumVector = list[float]
 type BinaryFunc_T = Callable[[Any, Any], Any]
 type UnaryFunc_T = Callable[[Any], Any]
